refactor(toc): decompose TableOfContents into focused per-concern files
Split the ~1000-line TableOfContents.tsx into a thin dispatcher plus
ten focused modules under src/components/toc/:

  types.ts                shared types, labels, helpers
  SectionEditForm.tsx     inline edit form
  SortableSectionItem.tsx single section row
  AddSectionItem.tsx      "+ add..." menu
  ChapterItem.tsx         single chapter row
  SectionList.tsx         SortableContext + items + empty state
  useSectionEdit.ts       inline-edit state hook
  useSectionDnd.ts        section drag/move/end + merge timer
  ArticleToc.tsx          article-mode body (DndContext + overlay)
  BookToc.tsx             book-mode body (chapter DndContext + overlay)

TableOfContents.tsx is now a 200-line dispatcher that owns only the
collapsed-state UI and the shared header, then routes to ArticleToc or
BookToc based on projectType.

No behavior change: same component tree, same handlers, same DnD
topology. Lays the groundwork for adding chapter content, multi-chapter
expansion, and cross-chapter section drag in subsequent commits.
